change types of hcn to int

+++ components/asic/include/bm1370.h
@@ -19,9 +19,9 @@ static const uint64_t BM1370_CORE_COUNT = 128;
 static const uint64_t BM1370_SMALL_CORE_COUNT = 2040;
 static const float BM1370_NONCE_PERCENT = BM1370_CORE_COUNT/256.0;  //maximum nonce space with 1 chip
 static const float BM1370_MIDSTATE_ENGINES = 16.0;
-static const float BM1370_HCN = 7861.0;
+static const int BM1370_HCN = 7861;
 static const float BM1370_HCN_MAX = 430000.0;
-static const float BM1370_HCN_PERCENT = BM1370_HCN/BM1370_HCN_MAX;       //hcn limit effect
+static const float BM1370_HCN_PERCENT = (float)BM1370_HCN/BM1370_HCN_MAX;       //hcn limit effect
 static const float BM1370_VERSION_PERCENT = 1.0;                    //version scan percent
 static const float BM1370_FULLSCAN_PERCENT = 0.5;                   //normalised value 1 means do the maximum space, 0 dont wait before sending work
